+++ exportToColleague/ncc/netconf_console/__init__.py
@@ -1,5 +1,5 @@
 """netconf-console2 application package."""
 
 __all__ = [
-    "ncc", "operations", "session", "trace", "config", "history", "namespaces"
+    "ncc", "operations", "session", "trace", "config", "history", "namespaces", "xmloutput"
 ]